feat: Operator IP management modified to be managed from the head office

+++ app/Http/Controllers/Manager/Service/OperatorIPController.php
@@ -8,8 +8,10 @@ use App\Http\Traits\ManagerTrait;
 use App\Http\Traits\ExtendResponseTrait;
 use App\Http\Traits\Models\EncryptDataTrait;
 
+use App\Enums\AuthLoginCode;
 use App\Http\Controllers\Ablilty\Ablilty;
 use App\Http\Controllers\Ablilty\EditAbleWorkTime;
+use App\Http\Controllers\Message\MessageController;
 
 use App\Http\Requests\Manager\Service\OperatorIPRequest;
 use App\Http\Requests\Manager\IndexRequest;
@@ -33,15 +35,17 @@ class OperatorIPController extends Controller
 
     /**
      * 목록출력
-     *
-     * @queryParam search string 검색어(제목)
      */
     public function index(IndexRequest $request)
     {
-        $search = $request->input('search', '');
-        $query = $this->operator_ips->where('brand_id', $request->user()->brand_id);
-        $data  = $this->getIndexData($request, $query);
-        return $this->response(0, $data);
+        if($request->user()->level >= 40)
+        {
+            $query = $this->operator_ips->where('brand_id', $request->user()->brand_id);
+            $data  = $this->getIndexData($request, $query);
+            return $this->response(0, $data);    
+        }
+        else
+            return $this->response(951);
     }
 
     /**
@@ -52,14 +56,24 @@ class OperatorIPController extends Controller
     {
         if(EditAbleWorkTime::validate() === false)
             return $this->extendResponse(1500, '지금은 작업할 수 없습니다.');
-    
-        $data = $request->data();
-        $res = $this->operator_ips->create($data);
-        $ips = $this->operator_ips->where('brand_id', $data['brand_id'])->get()->pluck('enable_ip')->all();
-        
-        AuthOperatorIP::set($data['brand_id'], $ips);
 
-        return $this->response($res ? 1 : 990, ['id'=>$res->id, 'brand_id'=>$data['brand_id']]);
+        [$result, $msg, $datas] = MessageController::operatorPhoneValidate($request);
+        if($result === AuthLoginCode::SUCCESS->value)
+        {
+            $data = $request->data();
+            $data['brand_id'] = $request->user()->brand_id;
+    
+            $res = $this->operator_ips->create($data);
+            $ips = $this->operator_ips->where('brand_id', $data['brand_id'])
+                ->get()->pluck('enable_ip')->all();
+            AuthOperatorIP::set($data['brand_id'], $ips);
+    
+            return $this->response($res ? 1 : 990, [
+                'id' => $res->id, 
+            ]);            
+        }
+        else
+            return $this->extendResponse($result, $msg, $datas);
     }
 
     /**
@@ -70,7 +84,7 @@ class OperatorIPController extends Controller
      */
     public function show(int $id)
     {
-        
+
     }
 
     /**
@@ -81,16 +95,30 @@ class OperatorIPController extends Controller
      */
     public function update(OperatorIPRequest $request, int $id)
     {
+        $operator_ip = $this->operator_ips->where('id', $id)->first();
+
         if(EditAbleWorkTime::validate() === false)
             return $this->extendResponse(1500, '지금은 작업할 수 없습니다.');
-        $data = $request->data();
-        $data = $this->setEncryptPersonalInfo($data);
-        $res  = $this->operator_ips->where('id', $id)->update($data);
-
-        $ips = $this->operator_ips->where('brand_id', $data['brand_id'])->get()->pluck('enable_ip')->all();
-        AuthOperatorIP::set($data['brand_id'], $ips);
-
-        return $this->response($res ? 1 : 990, ['id'=>$id, 'brand_id'=>$data['brand_id']]);
+        if(Ablilty::isBrandCheck($request, $operator_ip->brand_id) === false)
+            return $this->response(951);
+        else
+        {
+            [$result, $msg, $datas] = MessageController::operatorPhoneValidate($request);
+            if($result === AuthLoginCode::SUCCESS->value)
+            {
+                $operator_ip->enable_ip = $request->enable_ip;
+                $operator_ip->save();
+        
+                $ips = $this->operator_ips
+                    ->where('brand_id', $request->user()->brand_id)
+                    ->get()->pluck('enable_ip')->all();
+                AuthOperatorIP::set($operator_ip->brand_id, $ips);
+    
+                return $this->response(1, ['id' => $id]);
+            }
+            else
+                return $this->extendResponse($result, $msg, $datas);
+        }
     }
 
     /**
@@ -99,12 +127,31 @@ class OperatorIPController extends Controller
      * @urlParam id integer required PK
      * @return \Illuminate\Http\Response
      */
-    public function destroy(int $id)
+    public function destroy(Request $request, int $id)
     {
-        if(EditAbleWorkTime::validate() === false)
-            return $this->extendResponse(1500, '지금은 작업할 수 없습니다.');
-        $res = $this->operator_ips->where('id', $id)->delete();
-        return $this->response($res ? 1 : 990, ['id'=>$id]);
+        if($request->user()->level >= 40)
+        {
+            $operator_ip = $this->operator_ips->where('id', $id)->first();
+            if(EditAbleWorkTime::validate() === false)
+                return $this->extendResponse(1500, '지금은 작업할 수 없습니다.');
+            if(Ablilty::isBrandCheck($request, $operator_ip->brand_id) === false)
+                return $this->response(951);
+            else
+            {
+                [$result, $msg, $datas] = MessageController::operatorPhoneValidate($request);
+                if($result === AuthLoginCode::SUCCESS->value)
+                {
+                    $res = $this->operator_ips->where('id', $id)->delete();
+                    return $this->response($res ? 1 : 990, [
+                        'id' => $id
+                    ]);        
+                }
+                else
+                    return $this->extendResponse($result, $msg, $datas);
+            }
+        }
+        else
+            return $this->response(951);
     }
 
     static public function addIP($brand_id, $enable_ip)
